+++ prov/src/main/java/org/bouncycastle/jcajce/provider/asymmetric/ec/KeyAgreementSpi.java
@@ -158,6 +158,8 @@ public class KeyAgreementSpi
             };
         }
 
+        agreementCompleted();
+
         return null;
     }