+++ prov/src/main/java/org/bouncycastle/jcajce/provider/asymmetric/ecgost/KeyAgreementSpi.java
@@ -88,6 +88,8 @@ public class KeyAgreementSpi
             };
         }
 
+        agreementCompleted();
+
         return null;
     }