+++ prov/src/main/java/org/bouncycastle/jcajce/provider/asymmetric/ecgost12/KeyAgreementSpi.java
@@ -84,6 +84,8 @@ public class KeyAgreementSpi
             };
         }
 
+        agreementCompleted();
+
         return null;
     }