+++ prov/src/main/java/org/bouncycastle/jcajce/provider/asymmetric/sm9/KeyAgreementSpi.java
@@ -118,6 +118,8 @@ public class KeyAgreementSpi
             throw SecurityExceptions.invalidKeyException(e.getMessage(), e);
         }
 
+        agreementCompleted();
+
         return null;
     }